OceanGreensFunctionMethods

Documentation for OceanGreensFunctionMethods.

OceanGreensFunctionMethods.TracerInverseGaussian

OceanGreensFunctionMethods.abyssal_overturning

OceanGreensFunctionMethods.advective_diffusive_flux

OceanGreensFunctionMethods.advective_diffusive_flux

OceanGreensFunctionMethods.boundary_dimensions

OceanGreensFunctionMethods.boundary_flux

OceanGreensFunctionMethods.convergence

OceanGreensFunctionMethods.evolve_concentration

OceanGreensFunctionMethods.forcing_integrand

OceanGreensFunctionMethods.greens_function

OceanGreensFunctionMethods.integrate_forcing

OceanGreensFunctionMethods.intermediate_overturning

OceanGreensFunctionMethods.linear_probe

OceanGreensFunctionMethods.location_iodine129_history

OceanGreensFunctionMethods.location_transient_tracer_histories

OceanGreensFunctionMethods.mass

OceanGreensFunctionMethods.mass_convergence

OceanGreensFunctionMethods.model_dimensions

OceanGreensFunctionMethods.radioactive_decay

OceanGreensFunctionMethods.read_iodine129_history

OceanGreensFunctionMethods.read_transient_tracer_histories

OceanGreensFunctionMethods.steady_tracer_timeseries

OceanGreensFunctionMethods.timestep_initial_condition

OceanGreensFunctionMethods.tracer_point_source_history

OceanGreensFunctionMethods.tracer_source_history

OceanGreensFunctionMethods.tracer_tendency

OceanGreensFunctionMethods.tracer_tendency

OceanGreensFunctionMethods.tracer_tendency

OceanGreensFunctionMethods.tracer_timeseries

OceanGreensFunctionMethods.transient_tracer_timeseries

OceanGreensFunctionMethods.vertical_diffusion

OceanGreensFunctionMethods.TracerInverseGaussian — Type

TracerInverseGaussian(Γ,Δ)

using LinearAlgebra: NumberArray

The tracer inverse Gaussian distribution with mean Γ and width Δ has probability density function

\[G(𝐱, \tau) = \sqrt{\frac{\Gamma^3 }{4 \pi \Delta^2 \tau^3 }} \exp \left( - \frac{\Gamma (\tau - \Gamma)^2}{4 \Delta ^2 \tau}\right) \]

TracerInverseGaussian()              # Tracer Inverse Gaussian distribution with unit mean and unit width, i.e. TracerInverseGaussian(1, 1)
TracerInverseGaussian(Γ, Δ)          # Tracer Inverse Gaussian distribution with mean Γ and width Δ

params(d)           # Get the parameters, i.e. (Γ, Δ)
mean(d)             # Get the mean parameter, i.e. Γ
shape(d)            # Get the shape parameter, i.e. Δ

External links

Compare to Inverse Gaussian distribution on Wikipedia

OceanGreensFunctionMethods.abyssal_overturning — Method

abyssal_overturning(Ψ,model_dims)

Set volume flux, Ψ, in an abyssal overturning loop that satisfies the conservation of volume. Return a structure of Fluxes.

OceanGreensFunctionMethods.advective_diffusive_flux — Method

advective_diffusive_flux(C, Fv; ρ)

Advective-diffusive flux of tracer C given volume fluxes Fv and optional density ρ.

Arguments

C::DimArray: tracer distribution

Fv::DimArray: volume fluxes

ρ::Number=1035kg/m^3: uniform density

Returns

Fc::DimArray: tracer flux

OceanGreensFunctionMethods.advective_diffusive_flux — Method

advective_diffusive_flux(C, Fv; ρ)

Advective-diffusive flux of tracer C given volume fluxes Fv and optional density ρ.

Arguments

C::DimArray: tracer distribution

Fv::Fluxes: volume fluxes

ρ::Number=1035kg/m^3: uniform density

Returns

Fc::Fluxes: tracer flux

OceanGreensFunctionMethods.boundary_dimensions — Method

boundary_dimensions()

Define labels for the boundary's physical dimensions, as well as labels for the box names, consistently with the model dimensions. Use the format of DimensionalData.jl. Permits numerical quantities to be bundled with their meta-data. Dimensions are Unordered to avoid issues related to the alphabetical order.

OceanGreensFunctionMethods.boundary_flux — Method

boundary_flux(f::DimArray, C::DimArray, Fb::DimArray)

Convergence or net effect of boundary fluxes.

Arguments

f::DimArray: Dirichlet boundary condition

C::DimArray: tracer distribution

Fb::DimArray: boundary exchange volume flux

Returns

Jb::Fluxes: boundary tracer flux

OceanGreensFunctionMethods.convergence — Method

convergence(J)

Convergence of fluxes J of type Fluxes.

OceanGreensFunctionMethods.evolve_concentration — Method

evolve_concentration(C₀, A, B, tlist, source_history; halflife = nothing)

Integrate forcing vector over time to compute the concentration history. Find propagator by analytical expression using eigen-methods.

Arguments

C₀: initial tracer concentration

A: tracer transport information used in matrix calculations

B: boundary condition information used in matrix calculations

tlist: list of times to save tracer concentration

source_history::Function: returns Dirichlet boundary condition at a given time

halflife=nothing: radioactive half life (optional)

OceanGreensFunctionMethods.forcing_integrand — Method

forcing_integrand(t, tf, μ, V, B, source_history)

Integrand for boundary condition term in equation 10 (Haine et al., 2024).

Arguments

t: time

tf: final time

μ: eigenvalue diagonal matrix

V: eigenvector matrix

B: boundary condition matrix

source_history::Function: returns Dirichlet boundary condition at a given time

OceanGreensFunctionMethods.greens_function — Method

greens_function(τ,A)

Green's function for a box model (for steady transport given by the matrix 𝐀 for response at time t to a source at time t′ where τ = t - t′): the matrix exponential function of the elapsed time between the source time and field time.

𝐆(t) = eᴬᵗ

where 𝐆(t) is a N × N matrix with the spatial locations of field points (boxes) down its N rows and source points (boxes) along its N columns. Thus, the element 𝐆{i,j}(τ) quantifies transfer from a source at time t′ in box j to receiver at time t in box i.

OceanGreensFunctionMethods.integrate_forcing — Method

integrate_forcing(t0, tf, μ, V, B, source_history)

Integrate boundary condition term in equation 10 (Haine et al., 2024).

Arguments

t0: initial time

tf: final time

μ: eigenvalue diagonal matrix

V: eigenvector matrix

B: boundary condition matrix

source_history::Function: returns Dirichlet boundary condition at a given time

OceanGreensFunctionMethods.intermediate_overturning — Method

intermediate_overturning(Ψ,model_dims)

Set the volume flux, Ψ, in an intermediate overturning loop that satisfies the conservation of volume. Return a structure of Fluxes.

OceanGreensFunctionMethods.linear_probe — Method

linear_probe(funk, x, args...)

Probe a function to determine its linear response in matrix form. Assumes units are needed and available. A simpler function to handle cases without units would be nice.

Arguments

funk: function to be probed

x: input (independent) variable

halflife::Number: radioactive half life

args: the arguments that follow x in funk

Returns

A::DimArray{DimArray}: labeled transport information used in matrix operations

OceanGreensFunctionMethods.location_iodine129_history — Method

location_transient_tracer_histories()

URL of iodine-129 source history file.

OceanGreensFunctionMethods.location_transient_tracer_histories — Method

location_transient_tracer_histories()

URL of tracer source history file.

OceanGreensFunctionMethods.mass — Method

mass(V; ρ)

Seawater mass derived from the volume V and an optional input of density ρ.

OceanGreensFunctionMethods.mass_convergence — Method

mass_convergence(Fv)

Convergence of volume derived from a field of volume fluxes Fv, translated into a mass flux convergence with the assumption of uniform density.

OceanGreensFunctionMethods.model_dimensions — Method

model_dimensions()

Define labels for the model's physical dimensions, as well as labels for the box names. Use the format of DimensionalData.jl. Permits numerical quantities to be bundled with their meta-data. Dimensions are Unordered to avoid issues related to the alphabetical order.

OceanGreensFunctionMethods.radioactive_decay — Method

radioactive_decay(C, halflife)

Radioactive decay rate of tracer C with half life of halflife.

OceanGreensFunctionMethods.read_iodine129_history — Method

read_iodine129_history()

OceanGreensFunctionMethods.read_transient_tracer_histories — Method

read_transient_tracer_histories()

Read transient tracer source histories and save as a DimArray.

OceanGreensFunctionMethods.steady_tracer_timeseries — Method

steady_tracer_timeseries(tracername, A, B, halflife, tlist, mbox1, vbox1)

Simulate non-transient tracers and return tracer timeseries from one box.

Arguments

tracername: name of tracer

A: tracer transport matrix

B: boundary condition matrix

halflife: radioactive half life

BD: Dirichlet boundary condition

tlist: list of times to save tracer concentration

mbox: name of meridional box of interest

vbox: name of vertical box of interest

OceanGreensFunctionMethods.timestep_initial_condition — Method

timestepinitialcondition(C, μ, V, ti, tf)

Arguments

C::DimArray: tracer distribution at ti

μ: eigenvalue diagonal matrix

V: eigenvector matrix

ti: initial time

tf: final time

Returns

Cf::DimArray: tracer distribution at tf

OceanGreensFunctionMethods.tracer_point_source_history — Method

tracer_point_source_history(tracername, BD)

Return a function that yields transient tracer source history (such as CFCs) at given time.

Arguments

tracername: name of tracer in source history file

BD::DimArray: Dirichlet boundary condition compendium for many tracers

OceanGreensFunctionMethods.tracer_source_history — Function

tracer_source_history(t, tracername, box2_box1_ratio, BD = nothing)

Return source history values for all boundary points.

Arguments

t: time

tracername: name of tracer in source history file

box2_box1_ratio: ratio of boundary condition value in Mid-latitudes to High Latitudes

BD::DimArray=nothing: Dirichlet boundary condition compendium (optional)

OceanGreensFunctionMethods.tracer_tendency — Method

tracer_tendency(C)

Tracer tendency ∂C/∂t for the radioactive decay of a tracer C with half life halflife, for use with finding the radioactive contribution to a tracer transport matrix.

Arguments

C::DimArray: tracer distribution

halflife::Number: radioactive half life

Returns

dCdt::DimArray: tracer tendency

OceanGreensFunctionMethods.tracer_tendency — Method

tracer_tendency(C, f, Fv, Fb, V)

Tracer tendency ∂C/∂t for a tracer C, especially useful for finding a tracer transport matrix.

Arguments

C::DimArray: tracer distribution

f::DimArray: Dirichlet boundary condition

Fv::Fluxes: volume fluxes

Fb::Fluxes: volume fluxes

V::DimArray: box volume

Returns

dCdt::DimArray: tracer tendency

OceanGreensFunctionMethods.tracer_tendency — Method

tracer_tendency(f, C, Fv, Fb, V)

Tracer tendency ∂C/∂t for a boundary flux f, for use with finding B boundary matrix.

Arguments

f::DimArray: Dirichlet boundary condition

C::DimArray: tracer distribution

Fv::Fluxes: volume fluxes

Fb::Fluxes: volume fluxes

V::DimArray: box volume

Returns

dCdt::DimArray: tracer tendency

OceanGreensFunctionMethods.tracer_timeseries — Method

tracer_timeseries(tracername, A, B, tlist, mbox1, vbox1; BD=nothing, halflife=nothing)

Simulate tracers and return tracer timeseries from one box.

Arguments

tracername: name of tracer

A: tracer transport matrix

B: boundary condition matrix

tlist: list of times to save tracer concentration

mbox: name of meridional box of interest

vbox: name of vertical box of interest

BD=nothing: Dirichlet boundary condition

halflife=nothing: radioactive half life

OceanGreensFunctionMethods.transient_tracer_timeseries — Method

transient_tracer_timeseries(tracername, A, B, BD, tlist, mbox1, vbox1; halflife = nothing)

Simulate transient tracers and return tracer timeseries from one box.

Arguments

tracername: name of tracer

A: tracer transport matrix

B: boundary condition matrix

BD: Dirichlet boundary condition

tlist: list of times to save tracer concentration

mbox: name of meridional box of interest

vbox: name of vertical box of interest

halflife=nothing: radioactive half life

OceanGreensFunctionMethods.vertical_diffusion — Method

vertical_diffusion(Fv_exchange,model_dims)

Set vertical diffusive-like exchange flux Fv_exchange. Return a structure of Fluxes.